models.py
"""
Pydantic models for the Incident Triage Environment.

Defines the type-safe data structures used by the environment.
These are re-exported from openenv.core for spec compliance.
"""

from openenv.core.env_server.types import Action, Observation, State

__all__ = ["Action", "Observation", "State"]
