pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-incident-triage-env"
version = "0.1.0"
description = "Incident Triage Environment for OpenEnv - simulates production incident response for SRE agent training"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "fastmcp>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
server = "incident_triage_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["incident_triage_env", "incident_triage_env.server"]
package-dir = { "incident_triage_env" = ".", "incident_triage_env.server" = "server" }
